refactor(core): move tests to a separate file
Move the unit tests for the parsing logic in `anodized-core` to a
separate file, `crates/anodized-core/src/tests.rs`.
